backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
groq
python-dotenv
aiosqlite
sqlalchemy
